requirements.txt
stripe>=7.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.24.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.3
feedparser>=6.0.11
lxml>=5.1.0
email-validator>=2.0.0
sentry-sdk>=1.40.0
PyJWT>=2.8.0
